chore: Use NPM over Yarn
We want to be able to install @guardian/cdk from GitHub branches
as it makes testing easier.

GitHub does not have the compiled JS code and so we're relying on the
`prepare` lifecycle script in @guardian/cdk to run `tsc` to compile the TypeScript
files to JavaScript.

The `prepare` script:

> Runs on local npm install without any arguments
> NOTE: If a package being installed through git contains a prepare script, its dependencies and devDependencies will be installed, and the prepare script will be run, before the package is packaged and installed.

Sadly, it looks like Yarn v1 does not run the `prepare` script when
installing from GitHub, so let's move to NPM.

See:
  - https://github.com/yarnpkg/yarn/issues/5235#issuecomment-754993583
  - https://github.com/guardian/cdk/blob/7cad7e2bd3a9a604db777455dc4966d63cc950de/package.json#L18
  - https://docs.npmjs.com/cli/v6/using-npm/scripts#life-cycle-scripts