chore(deps): Rely on transitive AWS CDK libs
When updating `@guardian/cdk` we have to align the version of `aws-cdk`, `aws-cdk-lib`, and `constructs`
being used with those marked as peer dependencies from `@gaurdian/cdk`.

This makes updating quite complex as it's not just a matter of `npm update @guardian/cdk`.

In this change, we remove the explicit dependencies on these three libraries,
opting instead to rely on their (transitive) installation by NPM.

Since NPM v7, peer dependencies are automatically installed.

See:
 - https://github.blog/2020-10-13-presenting-v7-0-0-of-the-npm-cli/
 - `npm info @guardian/cdk@latest peerDependencies`